Fix a bug with getting valid class names.
Previously, if you called SomeFakeClass->new() inside of an eval{}
block, the execution wouldn't actually stop, and because the
%SomeFakeClass:: variable would have been magic'd into existence, the
call would register 'new' as being a valid key, and suddenly the class
would effectively be declared. In short, this means you could
accidentally create a class.

The new way can be circumvented by calling scalar(%SomeFakeClass::)
yourself, or simply saying %SomeFakeClass:: = (), but presumably if
you're doing that, you know what and why you're doing it, and it's
not just going to be something someone stumbles into.